requirements.txt
Flask>=2.3.0
groq>=0.4.0
python-dotenv>=1.0.0
bytez>=0.1.0
requests>=2.31.0

